% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.PCLasso.R
\name{cv.PCLasso}
\alias{cv.PCLasso}
\title{Cross-validation for \code{PCLasso}}
\usage{
cv.PCLasso(
  x,
  y,
  group,
  penalty = c("grLasso", "grMCP", "grSCAD", "gel", "cMCP"),
  nfolds = 5,
  standardize = TRUE,
  ...
)
}
\arguments{
\item{x}{A n x p design matrix of gene expression measurements with n samples and p genes, as in \code{PCLasso}.}

\item{y}{The time-to-event outcome, as a two-column matrix or \code{Surv} object, as in
\code{PCLasso}. The first column should be time on study (follow up time); the
second column should be a binary variable with 1 indicating that the event has
occurred and 0 indicating (right) censoring.}

\item{group}{A list of groups as in \code{PCLasso}. The feature (gene) names in \code{group}
should be consistent with the feature (gene) names in \code{x}.}

\item{penalty}{The penalty to be applied to the model. For group selection, one of grLasso,
grMCP, or grSCAD. For bi-level selection, one of gel or cMCP. See \code{grpsurv}
in the R package \code{grpreg} for details.}

\item{nfolds}{The number of cross-validation folds. Default is 5.}

\item{standardize}{Logical flag for \code{x} standardization, prior to fitting the model.
Default is \code{TRUE}.}

\item{...}{Arguments to be passed to \code{cv.grpsurv} in the R package \code{grpreg}.}
}
\value{
An object with S3 class "cv.PCLasso" containing:
\item{cv.fit }{An object of class "cv.grpsurv"}
\item{group.dt }{Groups with  features (genes) not included
in \code{x} being filtered out. }
}
\description{
Perform k-fold cross validations for the PCLasso model with grouped
covariates over a grid of values for the regularization parameter \code{lambda}.
}
\examples{
library("survival")

# load data
data(ExpMatrix)
data(survData)
data(PCGroup)

x = ExpMatrix
y = Surv(time=survData[,"time"], event=survData[,"status"])

# fit model
cv.fit1 <- cv.PCLasso(x, y, group = PCGroup, penalty = "grLasso", nfolds = 10)
}
