% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext2EntrezID.R
\name{ext2EntrezID}
\alias{ext2EntrezID}
\title{Transform selection features into EntrezID}
\usage{
ext2EntrezID(x)
}
\arguments{
\item{x}{Names of selected vars by \code{PCLasso} or \code{cv.PCLasso}.}
}
\value{
Selected genes.
}
\description{
Transform selection features into EntrezID at the gene level.
}
