% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext2Group.R
\name{ext2Group}
\alias{ext2Group}
\title{Convert selected features into ComplexID}
\usage{
ext2Group(x)
}
\arguments{
\item{x}{Names of selected vars by \code{PCLasso} or \code{cv.PCLasso}.}
}
\value{
Selected protein complexes.
}
\description{
Convert selected features into ComplexID at protein complex level
}
