% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GBM}
\alias{GBM}
\title{The GBM datasets}
\format{
A list with four GBM datasets:
\describe{
\item{GBM.train}{
A list with two items.
\code{Exp}: Expression data of \code{GBM.train}.
\code{survData}: Survival data of \code{GBM.train}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{GBM.test}{
A list with two items.
\code{Exp}: Expression data of \code{GBM.test}.
\code{survData}: Survival data of \code{GBM.test}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{GBM.valid1}{
A list with two items.
\code{Exp}: Expression data of \code{GBM.valid1}.
\code{survData}: Survival data of \code{GBM.valid1}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{GBM.valid2}{
A list with two items.
\code{Exp}: Expression data of \code{GBM.valid2}.
\code{survData}: Survival data of \code{GBM.valid2}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}
}
}
\usage{
data(GBM)
}
\description{
A dataset containing the GBM expression data and survival data.
}
\examples{
data(GBM)
}
\keyword{datasets}
