% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LUAD}
\alias{LUAD}
\title{The LUAD datasets}
\format{
A list with five LUAD datasets:
\describe{
\item{LUAD.train}{
A list with two items.
\code{Exp}: Expression data of \code{LUAD.train}.
\code{survData}: Survival data of \code{LUAD.train}. The first column is
the time on study (follow up time); the second column is a binary variable
with 1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{LUAD.test}{
A list with two items.
\code{Exp}: Expression data of \code{LUAD.test}.
\code{survData}: Survival data of \code{LUAD.test}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{LUAD.valid1}{
A list with two items.
\code{Exp}: Expression data of \code{LUAD.valid1}.
\code{survData}: Survival data of \code{LUAD.valid1}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{LUAD.valid2}{
A list with two items.
\code{Exp}: Expression data of \code{LUAD.valid2}.
\code{survData}: Survival data of \code{LUAD.valid2}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{LUAD.valid3}{
A list with two items.
\code{Exp}: Expression data of \code{LUAD.valid3}.
\code{survData}: Survival data of \code{LUAD.valid3}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}
}
}
\usage{
data(LUAD)
}
\description{
A dataset containing the LUAD expression data and survival data.
}
\examples{
data(LUAD)
}
\keyword{datasets}
