% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OV}
\alias{OV}
\title{The OV datasets}
\format{
A list with four OV datasets:
\describe{
\item{OV.train}{
A list with two items.
\code{Exp}: Expression data of \code{OV.train}.
\code{survData}: Survival data of \code{OV.train}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{OV.test}{
A list with two items.
\code{Exp}: Expression data of \code{OV.test}.
\code{survData}: Survival data of \code{OV.test}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{OV.valid1}{
A list with two items.
\code{Exp}: Expression data of \code{OV.valid1}.
\code{survData}: Survival data of \code{OV.valid1}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}

\item{OV.valid2}{
A list with two items.
\code{Exp}: Expression data of \code{OV.valid2}.
\code{survData}: Survival data of \code{OV.valid2}. The first column is the
time on study (follow up time); the second column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right)
censoring.}
}
}
\usage{
data(OV)
}
\description{
A dataset containing the OV expression data and survival data.
}
\examples{
data(OV)
}
\keyword{datasets}
