% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCGroup}
\alias{PCGroup}
\title{Protein complexes for "PCLasso"/"cv.PCLasso"}
\format{
A list of 2417 protein complexes.
}
\usage{
data(PCGroup)
}
\description{
A list of protein complexes. The genes in each protein complex are
represented by EntrezID, which are consistent with the gene names in the
expression data in \code{GBM}, \code{LUAD}, and \code{OV}.
}
\examples{
data(PCGroup)
}
\keyword{datasets}
