% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotKMCurve}
\alias{plotKMCurve}
\title{Plot Kaplan-Meier curve}
\usage{
plotKMCurve(
  rs,
  y,
  cutoff = median(rs),
  col.low = "#1F77B4",
  col.high = "#FF7F0E"
)
}
\arguments{
\item{rs}{Risk score of patients}

\item{y}{The time-to-event outcome, as a two-column matrix or \code{Surv}
object. The first column should be time on study (follow up time); the second
column should be a binary variable with 1 indicating that the event has
occurred and 0 indicating (right) censoring.}

\item{cutoff}{The cut-off that is used to divided the patients into two
groups according to \code{rs}.}

\item{col.low}{Color used to indicate low-risk patients.}

\item{col.high}{Color used to indicate high-risk patients.}
}
\description{
Plot Kaplan-Meier curve
}
\examples{
# load data
data(GBM)
data(PCGroup)

cv.fit1 <- cv.PCLasso(x = GBM$GBM.train$Exp,
                      y = GBM$GBM.train$survData,
                      group = PCGroup,
                      nfolds = 5)

# predict risk scores of samples in GBM.test
s <- predict(object = cv.fit1, x = GBM$GBM.test$Exp, type="link",
             lambda=cv.fit1$cv.fit$lambda.min)

plotKMCurve(rs = s, y = GBM$GBM.test$survData)

}
