% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRS}
\alias{plotRS}
\title{Plot risk scores of patients}
\usage{
plotRS(
  rs,
  y.status,
  cutoff = median(rs),
  col.dead = "#E64B35",
  col.alive = "#00A087",
  legend.position = c(0.2, 0.9)
)
}
\arguments{
\item{rs}{Risk score of patients}

\item{y.status}{A binary variable with 1 indicating that the event has
occurred and 0 indicating (right) censoring.}

\item{cutoff}{The cut-off that is used to divided the patients into two
groups according to \code{rs}.}

\item{col.dead}{The color that corresponds to 1 in \code{y.status}.}

\item{col.alive}{The color that corresponds to 0 in \code{y.status}.}

\item{legend.position}{Legend position.}
}
\description{
Plot risk scores of patients
}
\examples{
# load data
data(GBM)
data(PCGroup)

cv.fit1 <- cv.PCLasso(x = GBM$GBM.train$Exp,
                      y = GBM$GBM.train$survData,
                      group = PCGroup,
                      nfolds = 5)

# predict risk scores of samples in GBM.test
s <- predict(object = cv.fit1, x = GBM$GBM.test$Exp, type="link",
             lambda=cv.fit1$cv.fit$lambda.min)

plotRS(rs = s, y.status = GBM$GBM.test$survData$status)
}
