% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTDAUC}
\alias{plotTDAUC}
\title{Title}
\usage{
plotTDAUC(rs, y, npoint = 50, col = "red")
}
\arguments{
\item{rs}{Risk score of patients}

\item{y}{The time-to-event outcome, as a two-column matrix or \code{Surv}
object. The first column should be time on study (follow up time); the second
column should be a binary variable with 1 indicating that the event has
occurred and 0 indicating (right) censoring.}

\item{npoint}{Integer: number of time points for calculating time-dependent
AUC.}

\item{col}{Color of the time-dependent AUC curve.}
}
\description{
Title
}
\examples{
# load data
data(GBM)
data(PCGroup)

cv.fit1 <- cv.PCLasso(x = GBM$GBM.train$Exp,
                      y = GBM$GBM.train$survData,
                      group = PCGroup,
                      nfolds = 5)

# predict risk scores of samples in GBM.test
s <- predict(object = cv.fit1, x = GBM$GBM.test$Exp, type="link",
             lambda=cv.fit1$cv.fit$lambda.min)

plotTDAUC(rs = s, y = GBM$GBM.test$survData)
}
